Feature: Add Juiciness (#6)
* updated jump logic

updated jump logic so there isn't a ceiling and jump height is relative to the height the character jumped from.  Also increased gravity so jump feels less floaty when falling

* added relic open animation

* added player animation & updated scene w/ relic prefabs

Added jump, running, and falling animations along with a sprite flip when changing directions.  replaced old relic objects with the new relic prefabs that have the open animation

* updated fog of war clear to match main camera width

* changed isAlive to start as false to prevent flipping direction before clearing intro popup

* added camera shake on death

* fixed player animation not ending bug

* moved relics in front of player in sorting layers

* adjusted relic open animation time

* updated minimap cams to allow dynamic resolution

* added sound effects

Added death sound when the player dies, a box opening sound when picking up a relic, and a ta-da sound when reaching the finish

* updated readme w/ module 16 changes including new itch.io link